routh hurwitz table have been added for a given transfer function and polynomials
