+++ sympy/physics/control/__init__.py
@@ -1,6 +1,6 @@
 from .lti import (TransferFunction, Series, MIMOSeries, Parallel, MIMOParallel,
     Feedback, MIMOFeedback, TransferFunctionMatrix, StateSpace, gbt, bilinear, forward_diff,
-    backward_diff, phase_margin, gain_margin, routh_hurwitz_table)
+    backward_diff, phase_margin, gain_margin)
 from .control_plots import (pole_zero_numerical_data, pole_zero_plot, step_response_numerical_data,
     step_response_plot, impulse_response_numerical_data, impulse_response_plot, ramp_response_numerical_data,
     ramp_response_plot, bode_magnitude_numerical_data, bode_phase_numerical_data, bode_magnitude_plot,
@@ -9,8 +9,8 @@ from .control_plots import (pole_zero_numerical_data, pole_zero_plot, step_respo
 __all__ = ['TransferFunction', 'Series', 'MIMOSeries', 'Parallel',
     'MIMOParallel', 'Feedback', 'MIMOFeedback', 'TransferFunctionMatrix', 'StateSpace',
     'gbt', 'bilinear', 'forward_diff', 'backward_diff', 'phase_margin', 'gain_margin',
-    'routh_hurwitz_table', 'pole_zero_numerical_data', 'pole_zero_plot',
-    'step_response_numerical_data','step_response_plot', 'impulse_response_numerical_data',
-    'impulse_response_plot', 'ramp_response_numerical_data', 'ramp_response_plot',
+    'pole_zero_numerical_data', 'pole_zero_plot', 'step_response_numerical_data',
+    'step_response_plot', 'impulse_response_numerical_data', 'impulse_response_plot',
+    'ramp_response_numerical_data', 'ramp_response_plot',
     'bode_magnitude_numerical_data', 'bode_phase_numerical_data',
     'bode_magnitude_plot', 'bode_phase_plot', 'bode_plot']
